chore: organize components
