+++ packages/frontend/src/components/MainButton.tsx
@@ -1,4 +1,4 @@
-import Spinner from './Spinner';
+import Spinner from './global/Spinner';
 
 export const MainButton = (props: {
   color: string;
