+++ packages/frontend/src/components/post/NestedReply.tsx
@@ -2,7 +2,7 @@ import { useMemo } from 'react';
 import dayjs from 'dayjs';
 import relativeTime from 'dayjs/plugin/relativeTime';
 dayjs.extend(relativeTime);
-import { ButtonIcon } from './ButtonIcon';
+import { ButtonIcon } from '../ButtonIcon';
 import { IPostWithReplies, IRootPost } from '@/types/api';
 import { ReplyCount } from './ReplyCount';
 import { UserTag } from './UserTag';
