+++ packages/frontend/src/components/post/Post.tsx
