+++ packages/frontend/src/components/post/UserTag.tsx
@@ -13,7 +13,7 @@ export const UserTag = (props: UserTagProps) => {
     <div className="flex gap-2 items-center">
       <div className="flex gap-2 justify-center items-center">
         <Image alt={'profile'} src={imgURL ? imgURL : '/anon-noun.png'} width={30} height={30} />
-        <strong>{userId}</strong>
+        <p className="font-semibold">{userId}</p>
       </div>
       <p className="secondary">-</p>
       <p className="secondary">{date}</p>
