+++ packages/frontend/src/components/userInput/NewPost.tsx
